# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DocTrackr API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /smtp-settings:
    get:
      operationId: getSmtpSettings
      tags: [smtp]
      summary: Get SMTP settings
      responses:
        "200":
          description: SMTP settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmtpSettings"
    put:
      operationId: updateSmtpSettings
      tags: [smtp]
      summary: Update SMTP settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmtpSettingsInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmtpSettings"
  /smtp-settings/test:
    post:
      operationId: testSmtpSettings
      tags: [smtp]
      summary: Send test email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestEmailInput"
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  # Reminder Settings
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, all]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID with checklist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student with checklist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDetail"
        "404":
          description: Not found
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  # Employees
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, all]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID with checklist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee with checklist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDetail"
        "404":
          description: Not found
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  # Document Types
...
components:
  schemas:
...
    ActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    SmtpSettings:
      type: object
      properties:
        id:
          type: integer
        host:
          type: ["string", "null"]
        port:
          type: ["integer", "null"]
        username:
          type: ["string", "null"]
        fromEmail:
          type: ["string", "null"]
        fromName:
          type: ["string", "null"]
        secure:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - secure
        - createdAt
        - updatedAt
    SmtpSettingsInput:
      type: object
      properties:
        host:
          type: ["string", "null"]
        port:
          type: ["integer", "null"]
        username:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        fromEmail:
          type: ["string", "null"]
        fromName:
          type: ["string", "null"]
        secure:
          type: boolean
    TestEmailInput:
      type: object
      properties:
        toEmail:
          type: string
      required:
        - toEmail
...
    Student:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        classRoom:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        notes:
          type: ["string", "null"]
        parent1Name:
          type: string
        parent1Email:
          type: string
        parent1Phone:
          type: ["string", "null"]
        parent2Name:
          type: ["string", "null"]
        parent2Email:
          type: ["string", "null"]
        parent2Phone:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - fullName
        - status
        - parent1Name
        - parent1Email
        - createdAt
        - updatedAt
    StudentInput:
      type: object
      properties:
        fullName:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        classRoom:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        notes:
          type: ["string", "null"]
        parent1Name:
          type: string
        parent1Email:
          type: string
        parent1Phone:
          type: ["string", "null"]
        parent2Name:
          type: ["string", "null"]
        parent2Email:
          type: ["string", "null"]
        parent2Phone:
          type: ["string", "null"]
      required:
        - fullName
        - parent1Name
        - parent1Email
    StudentDetail:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        classRoom:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        notes:
          type: ["string", "null"]
        parent1Name:
          type: string
        parent1Email:
          type: string
        parent1Phone:
          type: ["string", "null"]
        parent2Name:
          type: ["string", "null"]
        parent2Email:
          type: ["string", "null"]
        parent2Phone:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        checklistItems:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
      required:
        - id
        - fullName
        - status
        - parent1Name
        - parent1Email
        - createdAt
        - updatedAt
        - checklistItems
    Employee:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        role:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - fullName
        - email
        - status
        - createdAt
        - updatedAt
    EmployeeInput:
      type: object
      properties:
        fullName:
          type: string
        role:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        notes:
          type: ["string", "null"]
      required:
        - fullName
        - email
    EmployeeDetail:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        role:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        checklistItems:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
      required:
        - id
        - fullName
        - email
        - status
        - createdAt
        - updatedAt
        - checklistItems
...
    ChecklistItem:
      type: object
      properties:
        id:
          type: integer
        personType:
          type: string
          enum: [student, employee]
        personId:
          type: integer
        personName:
          type: string
        personEmail:
          type: ["string", "null"]
        personStatus:
          type: string
          enum: [active, inactive]
        documentTypeId:
          type: integer
        documentTypeName:
          type: string
        documentTypeCategory:
          type: string
        templateFormUrl:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [missing, valid, expiring_soon, due_today, overdue]
        notes:
          type: ["string", "null"]
        lastReminderSentAt:
          type: ["string", "null"]
        nextReminderDueAt:
          type: ["string", "null"]
        daysUntilExpiry:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - personType
        - personId
        - personName
        - personStatus
        - documentTypeId
        - documentTypeName
        - documentTypeCategory
        - status
        - createdAt
        - updatedAt
...